+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.12.0]
+### Fixed
+- corrected the Latvian translation of the phrase "Select terminal"
+- fixed "Arrival email" additional service
+- added store email get from WP general info when no email is specified in plugin settings
+- disabled display of the return code in the delivery SMS
+- changed weight units to fixed value used in the plugin
+- added phone number service to courier-courier shipping
+- fixed label note variables
+
+### Improved
+- added a ability to disable Omniva shipping methods for specific shipping class
+- improved additional services working
+
 ## [1.11.1]
 ### Fixed
 - fixed error, when cart product dont have category (product category is "Uncategorized")